backend/requirements.txt
Flask
flask-mongoengine
jinja2==3.0.3
werkzeug
flask_api
langchain
chromadb
openai

# unstructured pdf loader
unstructured
pdf2image
pikepdf
pypdf
pdfminer.six
opencv-python
pandas
unstructured_pytesseract
unstructured_inference
# langchain
langchainhub
tiktoken
google-search-results